Moving GetFullEnvironment(..) to a new file LaunchConfigurationUtils.cs
